requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
mysql-connector-python==8.0.29
mysqlclient==2.1.0
protobuf==3.20.1
PyMySQL==1.0.2
python-dotenv==0.20.0
requests==2.27.1
urllib3==1.26.9
